Rework SRS Listener
